+++ .eslintrc.cjs
@@ -24,4 +24,4 @@ module.exports = {
     'no-new-object': 1, // 使用字面量创建对象 非new Object()
     'object-shorthand': 1, // 对象方法属性值缩写
   },
-}
+};