+++ src/preset/interceptors/index.ts
@@ -1,2 +1,3 @@
 export { default as commonCancelRequest } from './commonCancelRequest.js';
 export { default as commonThrottleRequest } from './commonThrottleRequest.js';
+export { default as commonParallelRequest } from './commonParallelRequest';